chore: group health,process,event api to system apis
